Allow each Capistrano Server to get it's own ec2 tags.
Because we sometimes add information that would be useful into custom
EC2 tags and it would be useful to refer to them during the deployment
process.

This adds an extra method onto the Capistrano Server objects that will
retrieve the AWS Instance tags from EC2 for a specific machine, and
return them as an AWS::ResourceTagCollection object directly.

This way you can write code in your recipes like

     on roles(:all) do |server|
        sudo :command, server.ec2_tags.fetch('command_args')
     end

Which is nice and useful.